Added submodule binSort
Added submodule binSort which takes a data array and a bins array as input and outputs an array containing the frequency of data within a given bin.
